ble: add barier for ble_disconnect
KRKNWK-16193
The ble_adapter_disconnect is a handler for application for sidewalk,
to request disconnect from current ble connection.
BLE layer has a callback that gets called when
state of the connection is changed.

In context of this function, we should wait for the proper callback
from BLE to be sure that we do not perform any other ble operation
during the deinitailization.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>