Build: fix linking order for Sidewalk libraries
Sidewalk libraries require gcc libraries for core functionaliteis.
Those requirements are provided by Zephyr.
Linker removes symbols that are not referenced, so if the
implemenetation ins provided in first library,
and second library uses the implementation.
Than by the time the linker starts to process the second library
the unused symbols are removed.

By changing the way Zephyr gets Sidewalk library, the order of linking
is fixded

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>