Build: Fix internal_west workflow
West complains that the manifest path can not be
one of the projects, even if this project is behind group filter.

This has been solved, by fetching sidewalk to internal_sidewalk,
and creating symbolic link to sidewalk directory.

this workflow affects only the automatic tools.
The customers that use NCS to fetch sidewalk, does not see any change.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>