samples: initial sketch for new template
KRKNWK-17953
[KRKNWK-17953]
This is sketch of new template that will simplify all sidewalk samples.
This commit shows the core structure that will be used by the samples.
The main idea is that the sample will be modular, and functionalities
will be controll by Kconfig options,
and different use cases will have separate prj.conf
to configure default set of features.

The main benefit is the common core of the sample,
the dts, Kconfig, thread structure, and handling applicaiton events

This is not completed feature, so the sample.yml
that defines samples for building and testing is not included yet

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>