shell: replace log messages in on dev cert to shell print
KRKNWK-18381
Because the shell is also responsible for handling log messages,
it is impossible to do so while handling the shell command.
This results in the delay between the creation of the log message
inside the shell handler and its process and printing.
Without this message the other side can not be sure
if the command succeded, so it can not proced with next command
wich causes significant delay.

Now the shell printer is used to pass the messages from
shell command and it does no go through logging system.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>