sample: fix issue with sending malformed payload on autoconnect
KRKNWK-18805
The issue has been investigated, and it was caused by the
`sid_put_msg` performing shallow copy of the data to send.
This was not intended, and will be fixed in next release.

Temporary fix for this issue, is to track the payload passed to
`sid_put_msg` and free it when message has been send, or when
there will be no further attempts with sending.

So this commit is temporary and should be reverted for the next release.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>