fix: concurrent read fo message list
In case a message has been scheduled, and we enter to DFU,
The application exits the sidewalk state,
and tries to clean the message list.
This was done in two steps, check if there are elements,
and if so, fetch them and clean.
But cleaning is also done in the message send callback from sidewalk.
Therefore it was possible that the sidewalk callback came between
the reading if the state is empty, and fetching the element,
and if the list had only one
the fetch will result in NULL pointer.

The issue has been resolved by fetching at one step,
and now only one context will get the element from the list.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>