assert: add error message to sid_pal_assert
Another assert has been discovered in the logs,
and it apears that it came from sid_assert.
There is no log for the location of the assert fail.
The CONFIG_ASSERT_NO_FILE_INFO is not set in the build
and CONFIG_ASSERT is enabled.
So the correct kernel assert has been called,
but it is not clear from the logs.

Added log message to make it more clear where issue happend.

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>