cmake: removing manual linking of libgcc.a
This commit removes the manual lookup and linking of libgcc.a.

Linking of C and runtime libraries are the responsibility of the higher
level build system which handles the toolchain and building of the
executable.

Current sidewalk implementation is broken in the sense that it fails to
consider the CPU architecture and just looks up any libgcc.a file.

For example the code might discover `arm-zephyr-eabi/12.2.0/libgcc.a`
when instead `arm-zephyr-eabi/12.2.0/thumb/v8-m.main+fp/hard/libgcc.a`
should be used, and thus causiong link errors such as:
> error: /.../12.2.0/libgcc.a(_arm_muldf3.o):
>                                    conflicting CPU architectures 17/2

and:
> error: zephyr/zephyr_pre0.elf uses VFP register arguments,
>        /.../12.2.0/libgcc.a(_fixunsdfdi.o) does not

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>