log: disable RTT backend when SHELL is in use.
Logs are send only to backends that are ready,
If some backend like SHELL takes some time to start,
and other backend is already initialized.
The messages will not be send to the uninitialized backend.
The solution is to disable the unused backednd
and therefore the logs will wait for the shell to initialize

Signed-off-by: Robert Gałat <robert.galat@nordicsemi.no>